+++ src/Filesystem.php
@@ -58,7 +58,6 @@ class Filesystem {
 		'mtime',
 		'size',
 		'touch',
-		'mkdir',
 		'rmdir',
 		'dirlist',
 	];
@@ -113,6 +112,59 @@ class Filesystem {
 		}
 
 		return call_user_func_array( [ $this->wp_filesystem, $method_name ], $arguments );
+
+	}
+
+	/**
+	 * Creates a directory.
+	 *
+	 * @throws \Exception If recursive parameter used width filesystem method other than direct.
+	 * @since [next]
+	 *
+	 * @param  string     $path      Path for new directory.
+	 * @param  int|false  $chmod     Optional. The permissions as octal number (or false to skip chmod).
+	 *                               Default false.
+	 * @param  string|int $chown     Optional. A user name or number (or false to skip chown).
+	 *                               Default false.
+	 * @param  string|int $chgrp     Optional. A group name or number (or false to skip chgrp).
+	 *                               Default false.
+	 * @param  bool       $recursive Whether to act recursively.
+	 * @return bool                  True on success, false on failure.
+	 */
+	public function mkdir( $path, $chmod = false, $chown = false, $chgrp = false, $recursive = false ) {
+
+		if ( ! $this->wp_filesystem instanceof \WP_Filesystem_Direct ) {
+			if ( $recursive ) {
+				throw new \Exception( 'Current filesystem method does not support recursive directory creation.' );
+			}
+
+			$this->wp_filesystem->mkdir( $path, $chmod, $chown, $chgrp );
+		}
+
+		// Safe mode fails with a trailing slash under certain PHP versions.
+		$path = untrailingslashit( $path );
+		if ( empty( $path ) ) {
+			return false;
+		}
+
+		if ( ! $chmod ) {
+			$chmod = FS_CHMOD_DIR;
+		}
+
+		// phpcs:ignore WordPress.PHP.NoSilencedErrors.Discouraged
+		if ( ! @mkdir( $this->path( $path ), $chmod, true ) ) {
+			return false;
+		}
+
+		if ( $chown ) {
+			$this->chown( $path, $chown );
+		}
+
+		if ( $chgrp ) {
+			$this->chgrp( $path, $chgrp );
+		}
+
+		return true;
 
 	}