+++ CHANGELOG.md
@@ -3,8 +3,11 @@ All notable changes to this project will be documented in this file.
 
 ## [next]
 
-Fixed:
-* Invalid method name causing warning
+### Added:
+* mkdir support for recursive dir creation when using Direct method.
+
+### Fixed:
+* Invalid method name causing warning.
 
 ## 1.0.0 - 28.01.2020